refactor(test): split cell_methods world/active_slot tests under cap
Tier 3 of #529 — bring two oversized test files under the 700-line hard
cap. Pure behavior-preserving mechanical moves; test logic unchanged.

player/world/tests.rs (1420) → tests/ directory:
  - mod.rs (93)            shared make_mgr_with_player fixture, dispatch
                           smoke test, region-trigger guard
  - auto_cycle.rs (638)    SET_AUTO_CYCLE toggle suite + quest-kill credit
  - reload.rs (320)        handle_reload core + Item_Reload/propId wire
  - reload_holster.rs (405) holster/phase choreography + in-combat guard

inventory/tests/active_slot_change.rs (1114) → two siblings:
  - active_slot_change.rs (601)    wire translation + holster choreography
  - active_slot_abilities.rs (526) auto-cycle stash + ability revoke/grant
                                   + onKnownAbilitiesUpdate broadcast

#[tokio::test] counts preserved: world 25→25, active_slot 12→12.
